+++ R/RcppExports.R
@@ -5,11 +5,11 @@ quad_index_cpp <- function(nx, ny, ydown) {
     .Call(`_textures_quad_index_cpp`, nx, ny, ydown)
 }
 
-quad_vertex_cpp <- function(nx, ny, ydown) {
-    .Call(`_textures_quad_vertex_cpp`, nx, ny, ydown)
+quad_vertex_cpp <- function(nx, ny, ydown, zh) {
+    .Call(`_textures_quad_vertex_cpp`, nx, ny, ydown, zh)
 }
 
-quad_vertex_matrix_cpp <- function(nx, ny, ydown) {
-    .Call(`_textures_quad_vertex_matrix_cpp`, nx, ny, ydown)
+quad_vertex_matrix_cpp <- function(nx, ny, ydown, zh) {
+    .Call(`_textures_quad_vertex_matrix_cpp`, nx, ny, ydown, zh)
 }