+++ src/RcppExports.cpp
@@ -19,36 +19,38 @@ BEGIN_RCPP
 END_RCPP
 }
 // quad_vertex_cpp
-NumericVector quad_vertex_cpp(IntegerVector nx, IntegerVector ny, LogicalVector ydown);
-RcppExport SEXP _textures_quad_vertex_cpp(SEXP nxSEXP, SEXP nySEXP, SEXP ydownSEXP) {
+NumericVector quad_vertex_cpp(IntegerVector nx, IntegerVector ny, LogicalVector ydown, LogicalVector zh);
+RcppExport SEXP _textures_quad_vertex_cpp(SEXP nxSEXP, SEXP nySEXP, SEXP ydownSEXP, SEXP zhSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< IntegerVector >::type nx(nxSEXP);
     Rcpp::traits::input_parameter< IntegerVector >::type ny(nySEXP);
     Rcpp::traits::input_parameter< LogicalVector >::type ydown(ydownSEXP);
-    rcpp_result_gen = Rcpp::wrap(quad_vertex_cpp(nx, ny, ydown));
+    Rcpp::traits::input_parameter< LogicalVector >::type zh(zhSEXP);
+    rcpp_result_gen = Rcpp::wrap(quad_vertex_cpp(nx, ny, ydown, zh));
     return rcpp_result_gen;
 END_RCPP
 }
 // quad_vertex_matrix_cpp
-NumericMatrix quad_vertex_matrix_cpp(IntegerVector nx, IntegerVector ny, LogicalVector ydown);
-RcppExport SEXP _textures_quad_vertex_matrix_cpp(SEXP nxSEXP, SEXP nySEXP, SEXP ydownSEXP) {
+NumericMatrix quad_vertex_matrix_cpp(IntegerVector nx, IntegerVector ny, LogicalVector ydown, LogicalVector zh);
+RcppExport SEXP _textures_quad_vertex_matrix_cpp(SEXP nxSEXP, SEXP nySEXP, SEXP ydownSEXP, SEXP zhSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< IntegerVector >::type nx(nxSEXP);
     Rcpp::traits::input_parameter< IntegerVector >::type ny(nySEXP);
     Rcpp::traits::input_parameter< LogicalVector >::type ydown(ydownSEXP);
-    rcpp_result_gen = Rcpp::wrap(quad_vertex_matrix_cpp(nx, ny, ydown));
+    Rcpp::traits::input_parameter< LogicalVector >::type zh(zhSEXP);
+    rcpp_result_gen = Rcpp::wrap(quad_vertex_matrix_cpp(nx, ny, ydown, zh));
     return rcpp_result_gen;
 END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
     {"_textures_quad_index_cpp", (DL_FUNC) &_textures_quad_index_cpp, 3},
-    {"_textures_quad_vertex_cpp", (DL_FUNC) &_textures_quad_vertex_cpp, 3},
-    {"_textures_quad_vertex_matrix_cpp", (DL_FUNC) &_textures_quad_vertex_matrix_cpp, 3},
+    {"_textures_quad_vertex_cpp", (DL_FUNC) &_textures_quad_vertex_cpp, 4},
+    {"_textures_quad_vertex_matrix_cpp", (DL_FUNC) &_textures_quad_vertex_matrix_cpp, 4},
     {NULL, NULL, 0}
 };