+++ src/quad-build.cpp
@@ -9,17 +9,15 @@ IntegerVector quad_index_cpp(IntegerVector nx, IntegerVector ny, LogicalVector y
 
 
 // [[Rcpp::export]]
-NumericVector quad_vertex_cpp(IntegerVector nx, IntegerVector ny, LogicalVector ydown) {
-  return quad::quad_vb(nx, ny, ydown);
+NumericVector quad_vertex_cpp(IntegerVector nx, IntegerVector ny,
+                              LogicalVector ydown,
+                              LogicalVector zh) {  // include z and homogeneous coord (0 and 1 respectively)
+  return quad::quad_vb(nx, ny, ydown, zh);
 }
 
 // [[Rcpp::export]]
-NumericMatrix quad_vertex_matrix_cpp(IntegerVector nx, IntegerVector ny, LogicalVector ydown) {
-  NumericVector v = quad_vertex_cpp(nx, ny, ydown);
-  // Set number of rows and columns to attribute dim
-  v.attr("dim") = Dimension(2, (nx[0] + 1) * (ny[0] + 1));
-
-  // Converting to Rcpp Matrix type
-  NumericMatrix quadvb_matrix = as<NumericMatrix>(v);
-  return quadvb_matrix;
+NumericMatrix quad_vertex_matrix_cpp(IntegerVector nx, IntegerVector ny,
+                                     LogicalVector ydown,
+                                     LogicalVector zh) {
+  return quad::quad_matrix_vb(nx, ny, ydown, zh);
 }